test(skills): the fuzz ratchet measured the mirror, not production
The seeded fence ratchet compared `_inside_fence`, a hand-written copy of the
production loop. A regression in production ORCHESTRATION left it green,
because it was not running production. That is not hypothetical: this same
mirror drifted earlier in this PR, omitting the malformed-closer re-scan, and
the baseline stayed green throughout.

`test_write_never_mutates_a_balanced_generated_document` drives the public
path, `find_fence_defects` and `repair_markdown_fences`, over all 6,000
generated documents and demands that a document the reference parser reads as
balanced comes back byte-identical.

Demonstrated rather than argued. Removing the `_container_closed` branch from
the repair loop:

  mirror-based ratchet      3 passed   (blind to it)
  production-driven test    3 failed   (catches it)

It is an absolute zero, not a baseline. Appending a fence to a well formed file
is the corruption this module exists to prevent, so there is no acceptable
non-zero count to ratchet down from. Documents carrying a mistaken closer are
excluded for the reason the shared module's Scope paragraph gives: repairing
those is the defect this tool exists to fix.

Splitting `test_fix_fences_contracts.py`, which the new test took to 507 lines
and the taste ratchet caught at 576 against 575. Raising a count baseline is
forbidden, so `TestVendoredInvocation` moves to test_fix_fences_vendoring.py:
it answers to the plugin-root invocation contract, not to the CommonMark
reference, which makes it a different file rather than a longer one. 376 and
155 lines, ratchet back at 575.

Reported by Copilot on PR #5322. 494 tests.

Refs #5322
